+++ .gitignore
@@ -1,0 +1,5 @@
+.idea
+/node_modules
+/dist
+.DS_Store
+.vscode
