+++ src/blocks/place-card/__image/_city/_karachaevsk/place-card__image_city_karachaevsk.css
