+++ src/blocks/popup/__picture/popup__picture.css
