+++ src/card.js
@@ -1,13 +1,5 @@
-class Card {
+export default class Card {
     constructor(name, link, openImagePopup, itemTemplate) {
-        /* Можно лучше: 
-    
-        Лучше передавать не отдельные параметры, а сразу весь объект с данными карточки,
-        т.к. представьте что у карточки появится ещё одно свойство (например author) которое нужно будет отобразить
-        Если у нас создание карточки вызывается как new Card(name,link), придется во всех местах
-        где вызывается создание карточки переписывать её вызов с new Card(name,link) на new Card(name, link, author) 
-        Если ли же мы передаем просто объект карточки в функцию ( new Card(cardData) ) нам придется гораздо меньше менять программу
-        */
         this.name = name;
         this.link = link;
         
@@ -41,19 +33,6 @@ class Card {
       this.cardElement.querySelector('.place-card__image').addEventListener('click', this.openPopup);
   }
 
-  // обрати внимание, как я задаю функцию обработчик,
-  // это нужно для того, что не потерялся контекст this
-  // в одном из тредов я кидал набор статей про контекст,
-  //  рекомендую почитать
-  /* Надо исправить:
-
-  При удалении карточки необходимо так же удалять обработчики событий с её элементов, это есть в чеклисте
-  "Если слушать события становится не нужно или элемент удаляется из разметки, обработчик должен быть удалён."
-  Для удаления обработчиков событий необходимо использовать метод removeEventListener. 
-
-  Удаление обработчиков собыйти с элементов лучше всего вынести в отдельный метод - removeEventListeners
-  и вызывать его в методе remove
-  */
   remove = (event) => {
     event.stopPropagation();
     this.removeEventListeners();
