+++ src/index.html
@@ -1,0 +1,80 @@
+<!DOCTYPE html>
+<html lang="ru">
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <meta http-equiv="X-UA-Compatible" content="ie=edge">
+  <title>Mesto</title>
+  <link rel="stylesheet" href="<%=htmlWebpackPlugin.files.css %>">
+</head>
+<body>
+  <div class="root">
+    <header class="header root__section">
+      <img src="<%=require('./images/logo.svg').default%>" alt="mesto logo" class="logo">
+    </header>
+    <div class="profile root__section">
+      <div class="user-info">
+        <div class="user-info__photo"></div>
+        <div class="user-info__data">
+          <h1 class="user-info__name"></h1>
+          <p class="user-info__job"></p>
+          <button class="button user-info__button-edit">Edit</button>
+        </div>
+        <button class="button user-info__button">+</button>
+      </div>
+    </div>
+    <div class="places-list root__section">
+
+      <!-- Пример вёрстки карточки: -->
+
+      <template class="item_template">
+        <div class="place-card">
+            <div class="place-card__image">
+                <button class="place-card__delete-icon"></button>
+            </div>
+            <div class="place-card__description">
+                <h3 class="place-card__name"></h3>
+                <button class="place-card__like-icon"></button>
+            </div>
+        </div>
+        </template>
+
+      <!-- конец примера. -->
+
+    </div>
+    <div class="popup">
+      <div class="popup__content">
+        <img src="<%=require('./images/close.svg').default%>" alt="" class="popup__close">
+        <h3 class="popup__title">Новое место</h3>
+        <form class="popup__form" name="new" novalidate>
+            <input type="text" id="namepicture" name="name" class="popup__input popup__input_type_name" placeholder="Название" minlength="2" maxlength="30" required>
+            <span  id="namepicture-error" class="popup__error"></span>
+            <input type="url" id="link" name="link" class="popup__input popup__input_type_link-url" placeholder="Ссылка на картинку" required>
+            <span  id="link-error" class="popup__error"></span>
+            <button type="submit" disabled class="button popup__button">+</button>
+        </form>
+      </div>
+    </div>
+    <div class="popup popup_user">
+      <div class="popup__content">
+        <img src="<%=require('./images/close.svg').default%>" alt="" class="popup__close popup__close_user">
+        <h3 class="popup__title">Редактировать профиль</h3>
+        <form class="popup__form popup__form_user" name="user" novalidate>  
+            <input type="text" id="name"  name="name" class="popup__input popup__input_type_name" placeholder="Имя"  minlength="2" maxlength="30" required>
+            <span  id="name-error" class="popup__error"></span>
+            <input type="text" id="job" name="job" class="popup__input popup__input_type_link-url" minlength="2" maxlength="30" required placeholder="О себе">
+            <span id="job-error" class="popup__error"></span>
+            <button type="submit" class="button popup__button popup__button_user popup__form_valid">Сохранить</button>
+        </form>
+      </div>
+    </div>
+    <div class="popup popup_image">
+      <div class="popup__container">
+        <img class="popup__picture">
+        <img src="<%=require('./images/close.svg').default%>" alt="" class="popup__close popup__close_image">
+      </div>
+    </div>
+  </div>
+  <script type="module" src="<%= htmlWebpackPlugin.files.js %>"></script>
+</body>
+</html>
